Fixed erroneous value in the agentmetrics.csv file
a value of 690 CCPOH was added to the database, this seems extremely unlikely as it would require someone to complete a call every 8 seconds.

this value messes with the graph and boosts team I's score significantly. I changed the value to 69 but it is still above what could be considered normal.